+++ frontend/src/workers/characterLoad.ts
